refactor: Add DB handling extensions
