--- AuthService/AuthService.IntegrationTests/IntegrationTestFactory.cs
@@ -1,91 +1,0 @@
-﻿using DotNet.Testcontainers.Builders;
-using DotNet.Testcontainers.Configurations;
-using DotNet.Testcontainers.Containers;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.AspNetCore.TestHost;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Npgsql;
-using System;
-using System.Collections;
-using System.Threading.Tasks;
-using Xunit;
-
-namespace AuthService.IntegrationTests
-{
-    public class IntegrationTestFactory<TProgram, TDbContext> : WebApplicationFactory<TProgram>, IAsyncLifetime
-    where TProgram : class where TDbContext : DbContext
-    {
-        private readonly TestcontainerDatabase _container;
-
-        public IntegrationTestFactory()
-        {
-            _container = new TestcontainersBuilder<PostgreSqlTestcontainer>()
-                .WithDatabase(new PostgreSqlTestcontainerConfiguration
-                {
-                    Database = "test_db",
-                    Username = "postgres",
-                    Password = "postgres",
-                })
-                .WithImage("postgres:11")
-                .WithCleanUp(true)
-                .Build();
-        }
-
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.ConfigureTestServices(services =>
-            {
-                services.RemoveDbContext<TDbContext>();
-                services.AddDbContext<TDbContext>(options => { options.UseNpgsql(_container.ConnectionString); });
-                services.EnsureDbCreated<TDbContext>();
-            });
-        }
-
-        public async Task InitializeAsync() => await _container.StartAsync();
-
-        public new async Task DisposeAsync() => await _container.DisposeAsync();
-
-        // insert vrednosti po potrebi
-        // genericki? -> trebaju nam imena kolona
-        // FAJL -> KSENOMIR
-
-        public long CountTableRows(string tableName)
-        {
-            long totalRows = -1;
-            using (var connection = new NpgsqlConnection(_container.ConnectionString))
-            {
-                using (var command = new NpgsqlCommand())
-                {
-                    connection.Open();
-                    command.Connection = connection;
-                    command.CommandText = "SELECT COUNT(*) FROM \"" + tableName + "\"";
-                    using (var reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            totalRows = (long)reader[0];
-                        }
-                    }
-                }
-            }
-            return totalRows;
-        }
-
-        public void DeleteById(string tableName, Guid id)
-        {
-            using (var connection = new NpgsqlConnection(_container.ConnectionString))
-            {
-                using (var command = new NpgsqlCommand())
-                {
-                    connection.Open();
-                    command.Connection = connection;
-                    command.CommandText = "DELETE FROM \"" + tableName + "\" WHERE \"Id\" = @id";
-                    command.Parameters.AddWithValue("@id", id);
-                    command.ExecuteNonQuery();
-                }
-            }
-        }
-    }
-}
