+++ AuthService/AuthService/Program.cs
@@ -44,4 +44,5 @@ app.MapControllers();
 app.UseMiddleware<ExceptionHandlerMiddleware>();
 
 app.Run();
+
 public partial class Program { }
